+++ CHANGELOG.md
@@ -6,6 +6,16 @@
 
 - Expose Replay and BrowserTracing on `@sentry/capacitor` ([#489](https://github.com/getsentry/sentry-capacitor/pull/489))
 
+### Fixes
+
+- Fixed sourcemaps for stacktraces other than localhost ([#494](https://github.com/getsentry/sentry-capacitor/pull/494))
+
+### Dependencies
+
+- Bump Sentry javascript 7.80.0 ([#499](https://github.com/getsentry/sentry-capacitor/pull/499))
+  - [changelog](https://github.com/getsentry/sentry-javascript/releases/tag/7.80.0)
+  - [diff](https://github.com/getsentry/sentry-javascript/compare/7.73.0...7.80.0)
+
 ## 0.14.0
 
 ### Features